requirements.txt
reportlab>=4.0.7
qrcode>=7.4.2
Pillow>=10.1.0
Faker>=20.1.0
typer>=0.9.0
lxml>=4.9.3
PyPDF2>=3.0.0